chore(line-chart-native): rename series loader file
